requirements.txt
python-decouple
mysqlclient
